Check convergence of REMA model

Access quick convergence checks from `TMB` and `nlminb`. Function modified from wham::check_convergence.

Usage

check_convergence(mod, ret = FALSE, f = "")

Arguments

mod

output from fit_rema

ret

T/F, return list? Default = FALSE, just prints to console

Value

a list with at least the first three of these components:

$convergence

From stats::nlminb, "0 indicates successful convergence for nlminb"

$maxgr

Max absolute gradient value, from `max(abs(mod$gr(mod$opt$par)))`

$maxgr_par

Name of parameter with max gradient

$is_sdrep

If TMB::sdreport was performed for this model, this indicates whether it performed without error

$na_sdrep

If TMB::sdreport was performed without error for this model, this indicates which (if any) components of the diagonal of the inverted hessian were returned as NA

See also

Examples

if (FALSE) { # \dontrun{
# placeholder for example
} # }