Plot the additional estimated observation error for biomass by strata and/or cpue by strata

Takes list output from tidy_rema and returns a list of ggplot2 objects to be plotted or saved.

Usage

plot_extra_cv(
  tidy_rema,
  save = FALSE,
  filetype = "png",
  path = NULL,
  xlab = NULL,
  biomass_ylab = "Biomass",
  cpue_ylab = "CPUE"
)

Arguments

tidy_rema

list out output from tidy_extra_cv, which includes inputs, model results, and confidence intervals for the total observation error (fixed + estimated)

save

(optional) logical (T/F) save figures as filetype in path. Default = FALSE. NOT YET IMPLEMENTED.

filetype

(optional) character string; type of figure file. Default = 'png'. NOT YET IMPLEMENTED.

path

(optional) directory path to location where figure files are to be saved if save = TRUE. NOT YET IMPLEMENTED.

xlab

(optional) label for x-axis of biomass and CPUE plots (e.g. 'Year'). Default = NULL.

biomass_ylab

(optional) label for y-axis of biomass plots (e.g. 'Biomass (t)'). Default = 'Biomass'.

cpue_ylab

(optional) label for y-axis of CPUE plots (e.g. 'Relative Population Number'). Default = 'CPUE'.

Value

a list of ggplot2 plots or character string messages about the data. Except for parameter estimates, the objects output from tidy_rema are the same outputted from this function.

See also

Examples

if (FALSE) { # \dontrun{
# placeholder for example
} # }