Read ADMB .rep file and return an R object of type 'list'

Code modified from original function provided by Steve Martell, D'Arcy N. Webber called by read_admb_re

Usage

read_rep(fn)

Arguments

fn

full path and name of ADMB output file to be read

Value

object of type "list" with ADMB outputs therein

Examples

if (FALSE) { # \dontrun{
read_rep(fn = 'inst/example_data/goasr.rep')
} # }